Move legacy class wrappers out of __init__.py
