+++ pyresample/bilinear/__init__.py
@@ -38,47 +38,16 @@ from ._numpy_resampler import (  # noqa: F401
     get_sample_from_bil_info,
     get_bil_info,
     NumpyBilinearResampler,
+    NumpyResamplerBilinear,
 )
 try:
     from .xarr import (  # noqa: F401
         XArrayBilinearResampler,
+        XArrayResamplerBilinear,
         CACHE_INDICES,
     )
 except ImportError:
     warnings.warn("XArray and/or zarr not found, XArrayBilinearResampler won't be available.")
     XArrayBilinearResampler = None
     CACHE_INDICES = None
-
-
-class XArrayResamplerBilinear(XArrayBilinearResampler):
-    """Wrapper for the old resampler class."""
-
-    def __init__(self, source_geo_def,
-                 target_geo_def,
-                 radius_of_influence,
-                 **kwargs):
-        """Initialize resampler."""
-        warnings.warn("Use of XArrayResamplerBilinear is deprecated, use XArrayBilinearResampler instead")
-
-        super(XArrayResamplerBilinear, self).__init__(
-            source_geo_def,
-            target_geo_def,
-            radius_of_influence,
-            **kwargs)
-
-
-class NumpyResamplerBilinear(NumpyBilinearResampler):
-    """Wrapper for the old resampler class."""
-
-    def __init__(self, source_geo_def,
-                 target_geo_def,
-                 radius_of_influence,
-                 **kwargs):
-        """Initialize resampler."""
-        warnings.warn("Use of NumpyResamplerBilinear is deprecated, use NumpyBilinearResampler instead")
-
-        super(NumpyResamplerBilinear, self).__init__(
-            source_geo_def,
-            target_geo_def,
-            radius_of_influence,
-            **kwargs)
+    XArrayResamplerBilinear = None
